Remove KSP from kotest-property extension
