--- extension/kotest-property-ksp/README.md
@@ -1,10 +1,0 @@
-# `kotlin-faker-kotest-property-ksp`
-
-[![Maven Central](https://img.shields.io/maven-central/v/io.github.serpro69/kotlin-faker-kotest-property-ksp?style=for-the-badge)](https://search.maven.org/artifact/io.github.serpro69/kotlin-faker-kotest-property-ksp)
-[![Sonatype Nexus (Snapshots)](https://img.shields.io/nexus/s/io.github.serpro69/kotlin-faker-kotest-property-ksp?label=snapshot-version&server=https%3A%2F%2Foss.sonatype.org&style=for-the-badge&color=yellow)](#downloading)
-
-Kotlin-faker [`kotest-property`](../kotest-property) and `kotest-property-ksp` artifacts provide faker-based [`Arb` generators](https://kotest.io/docs/proptest/property-test-generators.html) extensions via [KSP](https://kotlinlang.org/docs/ksp-overview.html) compiler plugin for [kotest property testing](https://kotest.io/docs/proptest/property-based-testing.html).
-
-## Usage
-
-Documentation for this extension is available at [serpro69.github.io/kotlin-faker/](https://serpro69.github.io/kotlin-faker/extensions/kotest-property-extension).
