--- extension/kotest-property-ksp/api/kotest-property-ksp.api
@@ -1,21 +1,0 @@
-public final class io/github/serpro69/kfaker/kotest/KotestArbProvider : com/google/devtools/ksp/processing/SymbolProcessorProvider {
-	public fun <init> ()V
-	public fun create (Lcom/google/devtools/ksp/processing/SymbolProcessorEnvironment;)Lcom/google/devtools/ksp/processing/SymbolProcessor;
-}
-
-public final class io/github/serpro69/kfaker/kotest/poet/KotlinPoetExtensionsKt {
-	public static final fun append (Lcom/squareup/kotlinpoet/ClassName;Ljava/lang/String;)Lcom/squareup/kotlinpoet/ClassName;
-	public static final fun asStringQuoted (Lcom/google/devtools/ksp/symbol/KSName;)Ljava/lang/String;
-	public static final fun getClassName (Lcom/google/devtools/ksp/symbol/KSDeclaration;)Lcom/squareup/kotlinpoet/ClassName;
-	public static final fun parameterizedWhenNotEmpty (Lcom/squareup/kotlinpoet/ClassName;Ljava/util/List;)Lcom/squareup/kotlinpoet/TypeName;
-	public static final fun writeTo (Lcom/squareup/kotlinpoet/FileSpec;Lcom/google/devtools/ksp/processing/CodeGenerator;)V
-}
-
-public final class io/github/serpro69/kfaker/kotest/poet/TypeNameKt {
-	public static final fun makeInvariant (Lcom/squareup/kotlinpoet/TypeVariableName;)Lcom/squareup/kotlinpoet/TypeVariableName;
-}
-
-public final class io/github/serpro69/kfaker/kotest/poet/TypeSpecKt {
-	public static final fun addClass (Lcom/squareup/kotlinpoet/FileSpec$Builder;Lcom/squareup/kotlinpoet/ClassName;Lkotlin/jvm/functions/Function1;)V
-}
-
