--- extension/kotest-property-ksp/build.gradle.kts
@@ -1,13 +1,0 @@
-plugins {
-    `faker-ext-conventions`
-}
-
-dependencies {
-    compileOnly(projects.extension.kotestProperty)
-    implementation(libs.ksp)
-    api(libs.bundles.kotlinpoet)
-}
-
-tasks.test {
-    useJUnitPlatform()
-}
