--- extension/kotest-property-ksp/src/main/resources/META-INF/services/com.google.devtools.ksp.processing.SymbolProcessorProvider
@@ -1,1 +1,0 @@
-io.github.serpro69.kfaker.kotest.KotestArbProvider
