--- extension/kotest-property/src/main/kotlin/io/github/serpro69/kfaker/kotest/FakerArb.kt
@@ -1,8 +1,0 @@
-package io.github.serpro69.kfaker.kotest
-
-import io.github.serpro69.kfaker.AbstractFaker
-import kotlin.reflect.KClass
-
-@Retention(AnnotationRetention.SOURCE)
-@Target(AnnotationTarget.PROPERTY, AnnotationTarget.FILE)
-annotation class FakerArb(vararg val fakers: KClass<out AbstractFaker>)
